+++ tests/conftest.py
@@ -4,6 +4,12 @@ import pytest
 import tempfile
 import posthog
 from unittest.mock import MagicMock
+from stat import S_IREAD, S_IRGRP, S_IROTH
+import platform
+
+if platform.system() == "Windows":
+    import win32security
+    import ntsecuritycon as con
 
 
 @pytest.fixture()
@@ -11,6 +17,45 @@ def tmp_directory():
     old = os.getcwd()
     tmp = tempfile.mkdtemp()
     os.chdir(str(tmp))
+
+    yield tmp
+
+    os.chdir(old)
+
+    # ignore unexpected permission error during test suite cleanup
+    shutil.rmtree(str(tmp), ignore_errors=True)
+
+
+@pytest.fixture()
+def tmp_readonly_directory():
+    old = os.getcwd()
+    tmp = tempfile.mkdtemp()
+    os.chdir(str(tmp))
+
+    # Read permissions
+
+    # TODO: Fix this for Windows
+    # Filesystem should not be writable, after running the below code
+    # But it is currently. Fix this
+    if platform.system() == "Windows":
+        # https://learn.microsoft.com/en-us/windows/win32/secauthz/well-known-sids
+        everyone = win32security.ConvertStringSidToSid("S-1-1-0")
+
+        # https://stackoverflow.com/questions/12168110
+        sd = win32security.GetFileSecurity(tmp, win32security.DACL_SECURITY_INFORMATION)
+        dacl = sd.GetSecurityDescriptorDacl()
+        dacl.AddAccessAllowedAce(
+            win32security.ACL_REVISION_DS,
+            con.FILE_GENERIC_READ,
+            everyone,
+        )
+        sd.SetSecurityDescriptorDacl(1, dacl, 0)
+        win32security.SetFileSecurity(tmp, win32security.DACL_SECURITY_INFORMATION, sd)
+    else:
+        # Although Windows supports chmod(), you can only set the file’s read-only flag
+        # with it(via the stat.S_IWRITE and stat.S_IREAD constants or
+        # a corresponding integer value). All other bits are ignored.
+        os.chmod(tmp, S_IREAD | S_IRGRP | S_IROTH)
 
     yield tmp