+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # CHANGELOG
 
 ## 0.2.13dev
+* [Fix] Fixed support for readonly file systems in telemetry (#63)
 * [Fix] Better validation when checking version updates ([#66](https://github.com/ploomber/core/issues/66))
+
 
 ## 0.2.12 (2023-06-06)